+++ CHANGELOG.md
@@ -1,8 +1,21 @@
-# Changelog
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## 0.1.4 - 2023-08-07
+### Added
+Default settings for plugin:
+```rust
+            config_webcam_device: 0,
+            config_webcam_width: 640,
+            config_webcam_height: 480,
+            config_webcam_framerate: 15,
+            config_webcam_autostart: true,
+            config_filter_type: SmoothingFilterType::LowPass(0.1),
+            config_filter_length: 10,
+```
+Added more informative exception handling
+### Changed
+Plugin now uses [camera_capture] instead of [rscam]. Should bring Windows support.
+config_webcam_device is now u32 type. 0 - default camera, or device number (Linux: number is added to "/dev/video{}", Windows: device number)
+### Removed
+Converter from yuyv not needed anymore
 
 ## 0.1.3 - 2023-08-05
 ### Added